havadurumu/requirements.txt
pyqt5
requests
matplotlib
python-dotenv